Updated diagnostics and added tests.
